pyproject.toml
[tool.poetry]
name = "fastapi-openai-sse-stream"
version = "0.1.0"
description = ""
authors = ["Tales Mousinho <tales.mousinho@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
uvicorn = "^0.27.0.post1"
fastapi = "^0.109.0"
sse-starlette = "^2.0.0"
openai = "^1.11.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
httpx = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
